+++ DuDoong-Domain/src/main/java/band/gosrock/domain/domains/order/domain/OrderLineItem.java
@@ -7,7 +7,6 @@ import band.gosrock.domain.common.vo.OptionAnswerVo;
 import band.gosrock.domain.common.vo.RefundInfoVo;
 import band.gosrock.domain.domains.cart.domain.CartLineItem;
 import band.gosrock.domain.domains.ticket_item.domain.TicketItem;
-import band.gosrock.domain.domains.ticket_item.domain.TicketType;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.CascadeType;
@@ -91,7 +90,7 @@ public class OrderLineItem extends BaseTimeEntity {
         return orderOptionAnswer.stream().map(OrderOptionAnswer::getOptionAnswerVo).toList();
     }
 
-    public Boolean isNeedPayment(){
+    public Boolean isNeedPayment() {
         return ticketItem.isNeedPayment();
     }
 }
