+++ DuDoong-Domain/src/main/java/band/gosrock/domain/domains/ticket_item/domain/TicketType.java
@@ -13,10 +13,8 @@ public enum TicketType {
     APPROVAL("APPROVAL");
     private String value;
 
-    /**
-     * 결제가 필요한지 상태를 반환하는 메서드
-     */
-    public Boolean isNeedPayment(){
+    /** 결제가 필요한지 상태를 반환하는 메서드 */
+    public Boolean isNeedPayment() {
         return this.equals(TicketType.FIRST_COME_FIRST_SERVED);
     }
 }
